Feat: review auth

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/ArticleDao.java
@@ -19,5 +19,5 @@ public interface ArticleDao {
 
     Long getMaxPage(String name, Long category, Long user, Long location);
 
-    List<Article> recommendedArticles(Integer articleId);
+    List<Article> recommendedArticles(Long articleId);
 }
